requirements.txt
#devops
black
click
pytest
pytest-cov
pylint
#rust based linter
ruff
boto3
#web
fastapi
uvicorn
#datascience
pandas
matplotlib
nbval
nbqa
#sql
mysql-connector-python
SQLAlchemy
pymysql